+++ src/main/java/co/orange/ddanzi/service/InterestService.java
@@ -3,9 +3,8 @@ package co.orange.ddanzi.service;
 import co.orange.ddanzi.domain.product.Product;
 import co.orange.ddanzi.domain.user.InterestProduct;
 import co.orange.ddanzi.domain.user.User;
-import co.orange.ddanzi.domain.user.pk.InterestProductId;
 import co.orange.ddanzi.dto.interest.InterestResponseDto;
-import co.orange.ddanzi.global.common.exception.Error;
+import co.orange.ddanzi.global.common.error.Error;
 import co.orange.ddanzi.global.common.response.ApiResponse;
 import co.orange.ddanzi.global.common.response.Success;
 import co.orange.ddanzi.repository.InterestProductRepository;
